Create currency.py
Function exchange(currency_from,currency_to,amount_from) need three parameters. The first is the upper abbreviation of the money to exchange from in string. The second parameter is the upper abbreviation of the money to exchange to in string and the third is the amount of money to exchange from in float. Then the function will return the result of money exchnged to according to the current exchange rate.