Create currency.py
Function exchange(currency_from,currency_to,amount_from) needs three parameters. The first is the abbreviation of money to exchange from in string, the second is the abbreviation of money to exchange to and the third is the amount of money to exchange from in float. Then the function will return the amount of money exchanged to in float according to the current exchange rate.