Update currency.py
The exchange rate is fixed at that of an exact time.
In this exchange, the user is changing amount_from money in currency currency_from to the currency currency_to. The value returned represents the amount in currency currency_to.
The value returned has type float.
Parameter currency_from: the currency on hand
 Precondition: currency_from is a string for a valid currency code
 Parameter currency_to: the currency to convert to
 Precondition: currency_to is a string for a valid currency code
 Parameter amount_from: amount of currency to convert
Precondition: amount_from is a float.
If the parameters input are not qualified to the preconditions, the result returned will be “Exchange currency code is invalid”
The test_exchange() function uses assert() function and does three tests.
The test_ALL() function can test all functions in this code. If the test is passed, it will print ‘test passed’.